Update install.sh
Don't blindly overwrite an existing .ssh/authorized_keys file (as
this will specifically lose ssh access to AWS instances).
Instead we have the keys we want to add in a separate file and 
compare the combined, sorted contents of extra file and existing file
with the installed one. We only update the installed file if there's
a difference.